Migrate vector to a separate module
