+++ same-alloc/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
 name = "same-alloc"
-version = "1.0.0-alpha"
+version = "1.0.0-beta"
 description = "Reuse an allocated buffer for data with different types."
 authors = ["Andreas Molzer <andreas.molzer@gmx.de>"]
-edition = "2018"
+edition = "2021"
 license = "MIT OR Apache-2.0 OR Zlib"
 documentation = "https://docs.rs/same-alloc"
 repository = "https://github.com/HeroicKatora/static-alloc"
